Add Timestampable protocol
