+++ AnnotatoSharedLibrary/Sources/AnnotatoSharedLibrary/Models/AnnotationPart.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-public protocol AnnotationPart: AnyObject, Codable {
+public protocol AnnotationPart: AnyObject, Codable, Timestampable {
     var id: UUID { get }
     var order: Int { get set }
     var height: Double { get set }
@@ -20,17 +20,5 @@ extension AnnotationPart {
 
     public func setHeight(to newHeight: Double) {
         self.height = newHeight
-    }
-
-    public func setCreatedAt(to createdAt: Date) {
-        self.createdAt = createdAt
-    }
-
-    public func setUpdatedAt(to updatedAt: Date) {
-        self.updatedAt = updatedAt
-    }
-
-    public func setDeletedAt(to deletedAt: Date) {
-        self.deletedAt = deletedAt
     }
 }
