+++ AnnotatoSharedLibrary/Sources/AnnotatoSharedLibrary/Models/AnnotationText.swift
@@ -16,10 +16,6 @@ public final class AnnotationText: Codable, AnnotationPart {
         content.isEmpty
     }
 
-    public var isDeleted: Bool {
-        deletedAt != nil
-    }
-
     @Published public private(set) var isRemoved = false
 
     public required init(
